+++ spec/forms/suite_import_form_spec.rb
@@ -12,24 +12,27 @@ RSpec.describe SuiteImportForm do
     before { stub_suite_create }
 
     it 'creates the suites' do
+      # from the fixture
+      count = 3
       described_class.import(file: file, building: building)
-      expect(Suite).to have_received(:create!).twice
+      expect(Suite).to have_received(:create!).exactly(count)
     end
     it 'returns nil in :redirect_object' do
       result = described_class.import(file: file, building: building)
       expect(result[:redirect_object]).to be_nil
     end
     it 'returns a success flash' do
+      # from the fixture
+      count = 3
       result = described_class.import(file: file, building: building)
-      expect(result[:msg].keys).to eq([:success])
+      expect(result[:msg][:success]).to match(/imported #{count} suites\./)
     end
     it 'creates the rooms' do
       # from the fixture
-      count = 6
+      count = 8
       described_class.import(file: file, building: building)
       expect(Room).to have_received(:create!).exactly(count)
     end
-
     def stub_suite_create
       allow(Suite).to receive(:create!)
         .and_return(instance_spy('Suite', number: 'L01'))
@@ -71,11 +74,16 @@ RSpec.describe SuiteImportForm do
       result = described_class.import(file: file, building: building)
       expect(result[:msg].keys).to match_array(%i(success error))
     end
+    it 'properly adds the failing number to the flash message' do
+      result = described_class.import(file: file, building: building)
+      expect(result[:msg][:error]).to match(/A32/)
+    end
     def stub_suite_create(building)
       allow(Suite).to receive(:create!)
-        .and_return(instance_spy('Suite', number: 'L01'))
-      allow(Suite).to receive(:create!).with(building: building, number: 'H42')
-                                       .and_raise('ERROR')
+        .and_return(instance_spy('Suite', number: 'A31'))
+      allow(Suite).to receive(:create!)
+        .with(building: building, number: 'A32', medical: true)
+        .and_raise(ActiveRecord::RecordInvalid)
       allow(Room).to receive(:create!)
     end
   end